boot: Memory mapping changes
linker.ld has been changed so that _bootloader_size is now the entire
image not just the .text section. This is because a copy of the e820 is
now being stored in .data, so needs to stay mapped.

main.c now copies the e820 to .data and passes a pointer to it to the
kernel's main function.