mem: Dual freelists + bitmap implementation
Physical memory is now tracked by a dual set of 2MB and 4kB freelists,
as well as a bitmap.

The bitmap is not currently fully implemented, but will eventually allow
coalescing multiple freed 4kB pages back into 2MB pages.

The E280 from the bootloader is parsed, and as many 2MB aligned pages
are carved out of it as possible, any unaligned/undersized areas are
added as 4kB pages.

In order to facilitate the freelist walking code, all physical memory is
now mapped into virtual memory. This is not ideal but it saved a lot of
headaches while the code is still immature.