VBE support
We now boot into 1024x768 and draw a background image, and overlay
the boot log.

Adds a font in unifont.c, and a bg image in ayame.c

tty.c is now the skeleton of a shared buffer for either text
or graphical modes to pull from and render.

At the moment, nothing is driving the screen to redraw, so
graphical output will 'hang' after the bg draws for the first time.